requirements.txt
pandas
tabulate
mysql-connector-python
